Preference refactoring
