+++ app/build.gradle
@@ -25,11 +25,12 @@ ext {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.google.code.gson:gson:2.3.1'
+    compile 'com.google.code.gson:gson:2.4'
     compile "com.android.support:appcompat-v7:${supportLibVersion}"
     compile "com.android.support:design:${supportLibVersion}"
     compile "com.android.support:recyclerview-v7:${supportLibVersion}"
     compile "com.android.support:support-v4:${supportLibVersion}"
+    compile "com.android.support:preference-v14:${supportLibVersion}"
     compile 'com.moparisthebest:junidecode:0.1.1'
     //compile 'com.ibm.icu:icu4j:55.1'
 }
