+++ app/src/main/AndroidManifest.xml
@@ -45,7 +45,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".SettingsActivity"
+            android:name=".settings.SettingsActivity"
             android:label="@string/title_activity_settings" />
 
         <service
