+++ app/src/main/java/com/althink/android/ossw/MainActivity.java
@@ -15,7 +15,6 @@ import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -26,6 +25,7 @@ import android.widget.Toast;
 import com.althink.android.ossw.plugins.PluginsFragment;
 import com.althink.android.ossw.service.OsswService;
 import com.althink.android.ossw.service.ble.BleConnectionStatus;
+import com.althink.android.ossw.settings.SettingsActivity;
 import com.althink.android.ossw.watchsets.ApplicationsFragment;
 import com.althink.android.ossw.watchsets.UtilitiesFragment;
 import com.althink.android.ossw.watchsets.WatchFacesFragment;
