+++ app/src/main/java/com/althink/android/ossw/settings/SettingsActivity.java
@@ -1,0 +1,84 @@
+package com.althink.android.ossw.settings;
+
+import android.app.FragmentManager;
+import android.app.FragmentTransaction;
+import android.os.Bundle;
+import android.preference.PreferenceManager;
+import android.support.v14.preference.PreferenceFragment;
+import android.support.v4.app.Fragment;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.preference.EditTextPreference;
+import android.support.v7.preference.ListPreference;
+import android.support.v7.preference.Preference;
+import android.support.v7.preference.PreferenceCategory;
+import android.support.v7.preference.PreferenceFragmentCompat;
+import android.support.v7.preference.PreferenceScreen;
+import android.support.v7.widget.Toolbar;
+import android.view.View;
+
+import com.althink.android.ossw.R;
+
+public class SettingsActivity extends AppCompatActivity implements
+        PreferenceFragment.OnPreferenceStartScreenCallback {
+
+    public static final String NOTIFICATION_VIBRATION_PREFIX = "notif_vibrate";
+    public static final String ALERT_VIBRATION_PREFIX = "alert_notif_vibrate";
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_settings);
+        Toolbar mToolbar = (Toolbar) findViewById(R.id.toolbar_actionbar);
+        mToolbar.setClickable(true);
+        setSupportActionBar(mToolbar);
+        getSupportActionBar().setHomeButtonEnabled(true);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+
+        mToolbar.setNavigationOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                onBackPressed();
+            }
+        });
+
+        if (savedInstanceState == null) {
+            PreferenceFragment fragment = new GeneralPreferenceFragment();
+            getFragmentManager().beginTransaction().replace(R.id.fragment_container, fragment, GeneralPreferenceFragment.FRAGMENT_TAG)
+                    .commit();
+        }
+    }
+
+    @Override
+    public boolean onPreferenceStartScreen(PreferenceFragment preferenceFragment,
+                                           PreferenceScreen preferenceScreen) {
+        FragmentTransaction ft = getFragmentManager().beginTransaction();
+        GeneralPreferenceFragment fragment = new GeneralPreferenceFragment();
+        Bundle args = new Bundle();
+        args.putString(PreferenceFragmentCompat.ARG_PREFERENCE_ROOT, preferenceScreen.getKey());
+        fragment.setArguments(args);
+        ft.add(R.id.fragment_container, fragment, preferenceScreen.getKey());
+        ft.addToBackStack(preferenceScreen.getKey());
+        ft.commit();
+        return true;
+    }
+
+    @Override
+    protected void onRestoreInstanceState(Bundle savedInstanceState) {
+        super.onRestoreInstanceState(savedInstanceState);
+    }
+
+    @Override
+    protected void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+    }
+
+    @Override
+    public void onBackPressed() {
+        int count = getFragmentManager().getBackStackEntryCount();
+        if (count == 0) {
+            super.onBackPressed();
+        } else {
+            getFragmentManager().popBackStack();
+        }
+    }
+}
