+++ app/src/main/res/values/strings.xml
@@ -58,7 +58,7 @@
 
     <string name="pref_title">Settings</string>
     <string name="select_font">Select font</string>
-    <string name="pref_notifications_title">Notification</string>
+    <string name="pref_notifications_title">Text notification</string>
     <string name="pref_notifications_desc">SMS, e-mail, missed call</string>
     <string name="pref_alert_notifications_title">Alert notification</string>
     <string name="pref_alert_notifications_desc">Incoming call, alarm clock</string>
