+++ app/src/main/res/values/template-styles.xml
@@ -32,6 +32,8 @@
         <item name="drawerArrowStyle">@style/DrawerArrowStyle</item>
 
         <item name="android:windowBackground">@color/myWindowBackground</item>
+
+        <item name="preferenceTheme">@style/PreferenceThemeOverlay.v14.Material</item>
      </style>
 
     <!-- ActionBar styles -->
