+++ app/src/main/res/xml/pref_general.xml
@@ -1,4 +1,7 @@
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+<PreferenceScreen
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:key="preference_screen"
+    android:title="@string/pref_title">
 
     <SwitchPreference
         android:defaultValue="true"
@@ -10,6 +13,7 @@
     <PreferenceCategory android:title="Notifications">
 
     <PreferenceScreen
+        android:key="notification_screen"
         android:title="@string/pref_notifications_title"
         android:summary="@string/pref_notifications_desc">
         <SwitchPreference
@@ -68,6 +72,7 @@
     </PreferenceScreen>
 
     <PreferenceScreen
+        android:key="alert_screen"
         android:title="@string/pref_alert_notifications_title"
         android:summary="@string/pref_alert_notifications_desc">
         <SwitchPreference
